+++ templates/cruds/foxx/widget/APP/main.js
@@ -18,26 +18,33 @@ const sessions = sessionsMiddleware({
 module.context.use(sessions);
 module.context.use(router);
 
+var typeCast = function(type, value) {
+  var value = unescape(value)
+  if(type == "integer") value = parseInt(value)
+  if(type == "float")   value = parseFloat(value)
+  return value
+}
+
 var fieldsToData = function(fields, body, headers) {
   var data = {}
   _.each(fields, function(f) {
     if(f.tr != true) {
       if(_.isArray(body[f.n])) {
-        data[f.n] = _.map(body[f.n], function(v) { return unescape(v) })
+        data[f.n] = _.map(body[f.n], function(v) { return typeCast(f.t,v) })
       } else {
         if(body[f.n] === undefined) {
           if(f.t == "boolean") data[f.n] = false
           else data[f.n] = null
         } else {
           if(f.t == "boolean") data[f.n] = true
-          else data[f.n] = unescape(body[f.n])
+          else data[f.n] = typeCast(f.t, body[f.n])
         }
       }
     } else {
       data[f.n] = {}
       if(_.isArray(body[f.n])) {
         data[f.n][headers['foxx-locale']] = _.map(
-          body[f.n], function(v) { return unescape(v) }
+          body[f.n], function(v) { return typeCast(f.t,v) }
         )
       } else {
         data[f.n][headers['foxx-locale']] = unescape(body[f.n])